+++ components/react/src/components/stories/card.stories.tsx
@@ -3,7 +3,6 @@ import { Stack } from 'styled-system/jsx'
 import { Button } from '~/components/ui/button'
 import { Card } from '~/components/ui/card'
 import { Field } from '~/components/ui/field'
-import { Input } from '~/components/ui/input'
 
 const meta: Meta = {
   title: 'Components/Card',
@@ -22,15 +21,11 @@ export const Base = () => {
         <Stack gap="4">
           <Field.Root>
             <Field.Label>Name</Field.Label>
-            <Field.Input asChild>
-              <Input placeholder="Name" />
-            </Field.Input>
+            <Field.Input placeholder="Name" />
           </Field.Root>
           <Field.Root>
             <Field.Label>Email</Field.Label>
-            <Field.Input asChild>
-              <Input type="email" placeholder="Email" />
-            </Field.Input>
+            <Field.Input type="email" placeholder="Email" />
           </Field.Root>
         </Stack>
       </Card.Body>