+++ components/solid/src/components/stories/field.stories.tsx
@@ -1,7 +1,5 @@
 import type { Meta } from 'storybook-solidjs'
 import { Field } from '~/components/ui/field'
-import { Input } from '~/components/ui/input'
-import { Textarea } from '~/components/ui/textarea'
 
 const meta: Meta = {
   title: 'Components/Field',
@@ -13,7 +11,7 @@ export const InputField = () => {
   return (
     <Field.Root>
       <Field.Label>Label</Field.Label>
-      <Field.Input asChild={(inputProps) => <Input {...inputProps()} />} />
+      <Field.Input />
       <Field.HelperText>Some additional Info</Field.HelperText>
       <Field.ErrorText>Error Info</Field.ErrorText>
     </Field.Root>
@@ -24,7 +22,7 @@ export const InputFieldInvalid = () => {
   return (
     <Field.Root invalid>
       <Field.Label>Label</Field.Label>
-      <Field.Input asChild={(inputProps) => <Input {...inputProps()} />} />
+      <Field.Input />
       <Field.ErrorText>Error Info</Field.ErrorText>
     </Field.Root>
   )
@@ -34,7 +32,7 @@ export const TextareaField = () => {
   return (
     <Field.Root>
       <Field.Label>Label</Field.Label>
-      <Field.Textarea asChild={(textareaProps) => <Textarea {...textareaProps()} />} />
+      <Field.Textarea />
       <Field.HelperText>Some additional Info</Field.HelperText>
     </Field.Root>
   )
@@ -44,7 +42,7 @@ export const TextareaFieldInvalid = () => {
   return (
     <Field.Root invalid>
       <Field.Label>Label</Field.Label>
-      <Field.Textarea asChild={(textareaProps) => <Textarea {...textareaProps()} />} />
+      <Field.Textarea />
       <Field.ErrorText>Error Info</Field.ErrorText>
     </Field.Root>
   )