+++ components/solid/src/components/ui/styled/field.tsx
@@ -1,6 +1,7 @@
 import { type Assign, Field } from '@ark-ui/solid'
 import type { ComponentProps } from 'solid-js'
-import { type FieldVariantProps, field } from 'styled-system/recipes'
+import { styled } from 'styled-system/jsx'
+import { type FieldVariantProps, field, input, textarea } from 'styled-system/recipes'
 import type { HTMLStyledProps } from 'styled-system/types'
 import { createStyleContext } from './utils/create-style-context'
 
@@ -26,11 +27,6 @@ export const HelperText = withContext<Assign<HTMLStyledProps<'span'>, Field.Help
   'helperText',
 )
 
-export const Input = withContext<Assign<HTMLStyledProps<'input'>, Field.InputBaseProps>>(
-  Field.Input,
-  'input',
-)
-
 export const Label = withContext<Assign<HTMLStyledProps<'label'>, Field.LabelBaseProps>>(
   Field.Label,
   'label',
@@ -41,9 +37,10 @@ export const Select = withContext<Assign<HTMLStyledProps<'select'>, Field.Select
   'select',
 )
 
-export const Textarea = withContext<Assign<HTMLStyledProps<'textarea'>, Field.TextareaBaseProps>>(
-  Field.Textarea,
-  'textarea',
-)
+export type InputProps = ComponentProps<typeof Input>
+export const Input = styled(Field.Input, input)
+
+export type TextareaProps = ComponentProps<typeof Textarea>
+export const Textarea = styled(Field.Textarea, textarea)
 
 export { FieldContext as Context } from '@ark-ui/solid'